Factor out the scheme-agnostic multi-open core
Mirrors the off-circuit split. `verifier/multi_open.rs` takes everything from the
first squeezed challenge onwards, and `kzg.rs` keeps only the KZG-specific parts.
Pure refactor: the verifying key is byte-identical.
